backend/requirements.txt
fastapi
uvicorn
pymongo
mysql-connector-python
pandas
python-dotenv
langchain
langchain-groq
langchain-experimental
aiofiles
